Trait mw_common::driver::Encoder

pub trait Encoder<G: Game, W: PlidsMuxWriter>: Sized {
    type EncodeError: Error;

    // Required method
    fn encode(
        &mut self,
        writer: &mut W,
        event_q: &mut Vec<(Plids, G::OutEvent)>
    ) -> Result<(), Self::EncodeError>;
}

Abstract interface for serializing a Game’s OutEvents into protocol streams

Used for games hosted on a network server, where the event stream needs to be sent over a network transport protocol.

Required Associated Types

type EncodeError: Error

Required Methods

fn encode( &mut self, writer: &mut W, event_q: &mut Vec<(Plids, G::OutEvent)> ) -> Result<(), Self::EncodeError>

Takes a queue of game events with associated destination Plids, generates a binary stream to be sent to each PlayerId.

Implementations are free to mangle and mutate the queue in-place as they see fit (typically for optimization purposes). It should be cleared before returning.