Trait mw_common::driver::Game

pub trait Game: Sized {
    type OutEvent: Clone + Send + Sync + 'static;
    type SchedEvent: Eq + Hash + Send + Sync + 'static;
    type InputAction: Clone + Send + Sync + 'static;
    type InitData: Send + Sync + 'static;

    // Required methods
    fn init<H: Host<Self>>(&mut self, host: &mut H, initdata: Self::InitData);
    fn unsched<H: Host<Self>>(&mut self, host: &mut H, event: Self::SchedEvent);
    fn input<H: Host<Self>>(
        &mut self,
        host: &mut H,
        plid: PlayerId,
        action: Self::InputAction
    );
}

Abstract interface through which the Host communicates with the Game

The “Game” is an implementation of the mechanics for a specific game mode. For example: the multiplayer MineWars game (proprietary) or the simplified singleplayer minesweeper mode (open-source).

Required Associated Types

type OutEvent: Clone + Send + Sync + 'static

Anything that happened that needs to be communicated to a player

When the Host calls either Game::input_action or Game::unsched to drive the Game, it can call Host::msg to send output events. They will be broadcast to all player ids selected with Plids.

type SchedEvent: Eq + Hash + Send + Sync + 'static

For things that need to be triggered on a timeout

Game code calls Host::sched, passing a value and time instant. The host code should store the value along with a timer. When the time instant has passed, host code calls Game::unsched, passing the value that was stored back to the game code.

type InputAction: Clone + Send + Sync + 'static

For things that are triggered by player input

Host code should call Game::input_action, passing an appropriate value, whenever the player wishes to perform an action in the game.

type InitData: Send + Sync + 'static

Data passed at initialization to set up the game.

Required Methods

fn init<H: Host<Self>>(&mut self, host: &mut H, initdata: Self::InitData)

Called once, at the start, before anything else

fn unsched<H: Host<Self>>(&mut self, host: &mut H, event: Self::SchedEvent)

Trigger a timer-driven event in the game

fn input<H: Host<Self>>( &mut self, host: &mut H, plid: PlayerId, action: Self::InputAction )

Process a player input